fixed typo

+++ IB_analysis/copy_files.py
@@ -13,18 +13,20 @@ subjects.sort()
 
 
 for subject in subjects:
-    src = root_dir + subject + "/%s_singleTone1.csv" % subject
-    dst = target_dir + "/%s_singleTone1.csv" % subject
-    copyfile(src, dst)
-
-    src = root_dir + subject + "/%s_singlePress.csv" % subject
-    dst = target_dir + "/%s_singlePress.csv" % subject
-    copyfile(src, dst)
-
-    src = root_dir + subject + "/%s_actionTone1.csv" % subject
-    dst = target_dir + "/%s_actionTone1.csv" % subject
-    copyfile(src, dst)
-
-    src = root_dir + subject + "/%s_actionPress.csv" % subject
-    dst = target_dir + "/%s_actionPress.csv" % subject
-    copyfile(src, dst)
+    if os.path.isfile(root_dir + subject + "/%s_singleTone1.csv" % subject):
+        
+        src = root_dir + subject + "/%s_singleTone1.csv" % subject
+        dst = target_dir + "/%s_singleTone1.csv" % subject
+        copyfile(src, dst)
+    
+        src = root_dir + subject + "/%s_singlePress.csv" % subject
+        dst = target_dir + "/%s_singlePress.csv" % subject
+        copyfile(src, dst)
+    
+        src = root_dir + subject + "/%s_actionTone1.csv" % subject
+        dst = target_dir + "/%s_actionTone1.csv" % subject
+        copyfile(src, dst)
+    
+        src = root_dir + subject + "/%s_actionPress.csv" % subject
+        dst = target_dir + "/%s_actionPress.csv" % subject
+        copyfile(src, dst)
